backend/requirements.txt
Django==3.2.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
django-redis
djongo
pymongo==3.12.1